% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_sig}
\alias{get_sig}
\title{Seasonal power}
\usage{
get_sig(y, wt, sig = 0.05)
}
\arguments{
\item{y}{Response}

\item{wt}{time in cycle}

\item{sig}{significance level (defaults to 0.05)}
}
\description{
\code{get_sig} Get significance info for one simulation
}

